+++ docs/src/js/stores/pet.js
@@ -1,0 +1,10 @@
+import { Store } from '~/rext';
+
+export default Store({
+  storeId: 'PetUserStore',
+  data: [
+    { pet: 'Cats', households: 38, total: 93 },
+    { pet: 'Dogs', households: 45, total: 79 },
+    { pet: 'Fish', households: 13, total: 171 }
+  ]
+})